+++ protocol/daemons/bridge/client/types/constants/logger.go
@@ -1,0 +1,5 @@
+package constants
+
+const (
+	BridgeDaemonModuleName = "bridge-daemon"
+)
